Style profile milestones and visited station cards
