+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/src/index.css" />
     <title>Zufallstour 3000 Berlin</title>
    <script>
       const isLocal = ['localhost', '127.0.0.1'].includes(location.hostname)
